Lib/im_lib/requirements.txt
Appium-Python-Client==2.1.2
PyYAML==6.0